Add level 2 AI for Go and Gomoku

+++ src/main/java/refactor/server/entity/Action.java
@@ -3,8 +3,8 @@ package refactor.server.entity;
 //Possible actions at a certain period of a game
 public class Action {
     public Position position;
-    public Integer score;
-    Action(Position position, Integer score) {
+    public Double score;
+    Action(Position position, Double score) {
         this.position = position;
         this.score = score;
     }
